add extra str check in iter. moved tests to test_toolkits.py cleaned up imports.
